+++ changelog.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added register entries support for channel state  - Register range: `0x8C-0x93`, Slot: `0x02`
 - Added support for number of channels register: `0xF4` Slot: `0x01`
 - Added support for valves, including reading name, and showing a sensor for state.
+- Added support for changing valve state
 
 ### Changed
 - Updated documentation based on new understanding of header checksum